app/requirements.txt
click==7.1.2
Flask==2.0.3
jmespath==0.10.0
PyYAML==5.4.1